+++ cli/helpers/server_runner.go
@@ -14,6 +14,7 @@ type HooksServerRunConfig struct {
 	Env               []string
 	LogLevel          string
 	Production        bool
+	DebugBindAddress  string
 	Debug             bool
 	Output            *scriptrunner.OutputConfig
 }
@@ -36,7 +37,7 @@ func NewHooksServerRunner(log *zap.Logger, cfg *HooksServerRunConfig) *scriptrun
 	var scriptArgs []string
 
 	if cfg.Debug {
-		scriptArgs = append(scriptArgs, "--inspect")
+		scriptArgs = append(scriptArgs, "--inspect="+cfg.DebugBindAddress)
 	}
 	scriptArgs = append(scriptArgs, cfg.ServerScriptFile)