+++ packages/testsuite/apps/openapi/.wundergraph/wundergraph.config.ts
@@ -10,6 +10,32 @@ const notes = introspect.openApiV2({
 		filePath: './api.yaml',
 	},
 	baseURL: new EnvironmentVariable('OPENAPI_URL', 'http://localhost:8090/'),
+	schemaExtension: `
+		type Reminder {
+			id: Int
+			date: Int
+			time: Int
+			what: String
+		}
+
+		type JSONData  {
+			id: Int!
+			name: String!
+			date: Int!
+		}
+	`,
+	replaceCustomScalarTypeFields: [
+		{
+			entityName: 'noteByID_200_response',
+			fieldName: 'jsonData',
+			responseTypeReplacement: '[JSONData]',
+		},
+		{
+			entityName: 'Query',
+			fieldName: 'reminders',
+			responseTypeReplacement: '[Reminder]',
+		},
+	],
 });
 
 const unions = introspect.openApiV2({