+++ packages/swr/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.9](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.17.8...@wundergraph/swr@0.17.9) (2023-08-17)
+
+**Note:** Version bump only for package @wundergraph/swr
+
+## [0.17.8](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.17.7...@wundergraph/swr@0.17.8) (2023-08-14)
+
+**Note:** Version bump only for package @wundergraph/swr
+
+## [0.17.7](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.17.6...@wundergraph/swr@0.17.7) (2023-08-13)
+
+**Note:** Version bump only for package @wundergraph/swr
+
+## [0.17.6](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.17.5...@wundergraph/swr@0.17.6) (2023-08-03)
+
+**Note:** Version bump only for package @wundergraph/swr
+
+## [0.17.5](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.17.4...@wundergraph/swr@0.17.5) (2023-08-03)
+
+**Note:** Version bump only for package @wundergraph/swr
+
+## [0.17.4](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.17.3...@wundergraph/swr@0.17.4) (2023-08-02)
+
+**Note:** Version bump only for package @wundergraph/swr
+
+## [0.17.3](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.17.2...@wundergraph/swr@0.17.3) (2023-08-01)
+
+**Note:** Version bump only for package @wundergraph/swr
+
 ## [0.17.2](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.17.1...@wundergraph/swr@0.17.2) (2023-07-28)
 
 **Note:** Version bump only for package @wundergraph/swr